Create test-driven websocket client, server plugin
Having sort of figured out how to do things with Hapi in the
hapi-websocket-server and hapi-websocket-client spikes, it was time to
do some proper TDD.

I think this test is pretty cool. I can spin up a Hapi server, register
my implementation as a plugin, and then start up a client object in the
same test and exercise them together. I decided not to mock out the
server's domain layer, as the test was easier to reason about with a
real Game in play, which makes this a somewhat far-reaching integration
test: it integrates the frontend's client object, the backend server
layer, and the backend domain.

That could change in the future. If the domain gets complicated enough
that it gets hard to reason about this test, then test doubles could be
injected into the server plugin in its registration options to isolate
just the websocket stuff. But, the test in go-fish-websocket-client
would continue to exercise the client and server together, which I think
will negate the need for an end-to-end test.

I did find something troubling: if an error gets thrown in the server
code, or in the client's subscription handler, it appears to prevent the
client from disconnecting at the end of the test, which causes the test
to hang. I've added a universal try-catch in the subscription handler to
protect against that, but that means that an error that doesn't
specifically result in a failed assertion would still allow tests to
pass, which I don't love. I may raise an issue with @hapi/nes to see
what's up.

I still need to move the implementation code out to its proper homes.