Use heroku port, secure web socket url
Heroku requires you to start your application on an assigned port,
accessed via an environment variable.

Also, when deployed, browsers will refuse to connect to a ws://
websocket URL when the page is being served from an https:// url. My fix
for this is very hacky, and means that you cannot run the production
build of the frontend on a local machine. At some point I should find a
more sophisticated way to choose the proper websocket url.