Add support for custom player names to game engine
The game engine now assigns a unique identifier to a player when
`addPlayer()` is called, and `name` is an optional field in PlayerState.

To maintain compatibility with the rest of the system, the server plugin
treats the assigned player id as the player name that gets sent to the
front end. This makes the game look a little uglier, but it's still
playable.

The next step is to add support for updating the player name to the
server plugin and client, and after that, to dress it up nicely in the
UI.