Correct references from playerName to playerId
A player's name is no longer their unique identifier in the game engine.
In future commits I'll add UI support for updating the name (which is now
an optional field on the player information in the game state) and
displaying the name instead of the ID above the player's play area.