Route /game/:gameId paths to index.html
I thought this was already working, but it turned out the Create React
App dev server was deceiving me.

Debugging this required running the production-mode server on my local
machine, so I had to improve the hack for constructing the websocket URL
(previously it assumed that if NODE_ENV was production, wss could always
be used, but when running locally that's not the case).