Prevent refreshes from creating new players
The game now stores an association between game ids and player ids in
local storage, and if a player id for the current game id is found, is
uses that one instead of adding another to the game when joining.

This means that if people open a link more than once, or refresh their
browser window, we won't have abandoned players lying around.

Note that because the game id/player id map is stored in local storage,
it persists between windows! Opening new tabs or windows won't allow
players to join a game multiple times. This does make things a little
obnoxious for local development, where it's convenient to be able to use
multiple windows to add a bunch of players to a game for testing
purposes. Therefore, when process.env.NODE_DEV is set to development
(which the Create React App dev server does), local storage isn't used.

If you need to test something about the local storage logic, you can
bypass the dev server on port 3000 and look at localhost:5000, which has
the production-like behavior.

I found a frustrating gotcha while doing this, which is that if you
don't wait for the promise returned by gameplayClient.joinGame() to
resolve before doing other stuff (especially in the client's user
tests), it gets weird, and very easy to cause a socket disconnection.

I also decided to make game ids less predictable, to lessen the
likelihood of people accidentally crashing games they didn't intend to.