Restore game from local state after server bounce
Before this commit, the game will crash if the server restarts in the
middle of play, because without a backing database, all game data is
lost.

This commit solves this problem without introducing a database (which
would make it more costly for communities to run an instance, and more
complicated for them to set one up) by having clients detect when the
game has been lost and send a locally cached copy of the game state to
the server to replace it.

This is only a very basic treatment. I think it will fall over if
players try to take an action while the server is still down, or
immediately after it starts up again (while the Nes client is still
trying to reconnect).

I'm also not entirely happy with the test; I have a 2-second sleep after
the server bounce to give the client a chance to reconnect, which is
slow and brittle. The whole test is (necessarily?) pretty slow; I had to
increase the Jest time cap to make sure it had enough space to complete.

A more robust solution would be to use the onDisconnect and onConnect
handlers on the Nes client (and the willReconnect parameter passed to
the onDisconnect handler) to work out whether the client is in the
process of reconnecting, and respond appropriately.

However, the test will be a little tricky, as on my first pass, I found
that the onDisconnect trigger can happen significantly after the
server.stop() promise resolves, so getting the assertions to run at the
right time will require some finesse.

I may also want to add some messaging to the user to alert them when the
server is down, so they know to wait a few seconds to see if it comes
back up.