Refactor browser client and spec
This commit removes the sleep statement introduced previously in the
test for recovering from a restarted server. The test now waits (up to 3
seconds, via the "eventually" spec helper method) for the client to
recover its connection before proceeding, instead of arbitrarily waiting
2 seconds. (This makes the test less brittle, and also almost a
second faster most of the time.)

To make that work I did have to add an "isConnected" method onto the
client's public interface, but I don't mind that. It could conceivably
be useful if I end up wanting to communicate to the user when the
connection to the server is down. In fact, the fake client used in the
UI tests had already extended the real interface with an "isConnected"
method for test-only use there.